Make make_scalar_function() result candidate for inlining
`make_scalar_function` serves as a template to implement functions,
abstracting away processing of scalar values.  While
`make_scalar_function` is a template function, it's not eligible for
effective inlining because of the `Arc<Fn>` return type.

This commit removes `Arc` and unlocks inlining. The change impact was
verified manually using on a similar `make_scalar_function` helper in
DataFusion.

The change saves some small overhead on each invocation of a UDF, but
doesn't improve per-row performance.